% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\name{cdfA}
\alias{cdfA}
\alias{cdfB}
\title{cdfA and cdfB}
\usage{
cdfA(v)

cdfB(v)
}
\description{
Empirical cumulative distribution functions \code{\link{ecdf}} describing
DNA-concentration-dependent error rates based on data for \emph{HLA-A} and
\emph{HLA-B} extracted for the period between 01/09/2014 and 23/05/2015.
}
\examples{
data(cdfA)
plot(cdfA)

## extract the number of observations underlying an ecdf
get("nobs", environment(cdfA))
}

