% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.r
\name{clean_hla_data}
\alias{clean_hla_data}
\title{Remove low resolution four-digit codes and cases of unknown alleles}
\usage{
clean_hla_data(x)
}
\arguments{
\item{x}{A \code{\link{HLA}} object}
}
\description{
Remove low resolution four-digit codes and cases of unknown alleles
}

