% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{concentration}
\alias{concentration}
\title{Template DNA concentrations}
\format{Data tables with 4 columns:
\describe{
 \item{lims_donor_id}{Sample identifier.}
 \item{ymd}{Date of measurement}
 \item{conc}{DNA concentration in ng/µl}
 \item{provenance}{Origin of sample. One of "DE", "PL", or "UK"}
}}
\usage{
concentration
}
\description{
DNA concentration measures for 24150, 24304, and 25049 random samples
from Germany, Poland, and the UK, respectively.
}
\examples{
  concentration
}
\keyword{datasets}

