% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.r
\name{disambiguate_codes}
\alias{disambiguate_codes}
\title{Disambiguate NMDP and G codes}
\usage{
disambiguate_codes(x, gtbl, ntbl4)
}
\arguments{
\item{x}{An allele}
}
\description{
Disambiguate NMDP and G codes
}
\details{
Expect gtbl and ntbl4 (4-field subtypes)
in the parent environment
}
\keyword{internal}

