% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna-concentration.r
\name{dkms_dna_concentration}
\alias{dkms_dna_concentration}
\title{Extract DNA concentration for LIMS_DONOR_ID from NGSREP}
\usage{
dkms_dna_concentration(lims_donor_id, gene, chunksize = 1000, ncores = 1)
}
\arguments{
\item{lims_donor_id}{LIMS_DONOR_ID}

\item{gene}{One of \sQuote{A}, \sQuote{B}, \sQuote{C},
\sQuote{DRB1}, \sQuote{DQB1}, or \sQuote{DPB1}.}

\item{chunksize}{Submit queries in chunks of LIMS_DONOR_IDs.}

\item{ncores}{Parallelise queries.}
}
\value{
A table with fields:
\itemize{
 \item{\code{lims_donor_id}}
 \item{\code{ymd}}
 \item{\code{conc}}
 \item{\code{provenance}}
}
}
\description{
Extract DNA concentration for LIMS_DONOR_ID from NGSREP
}
\keyword{internal}

