% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generics.r
\name{dna_version_id}
\alias{dna_version_id}
\title{Extract the DNA Version ID}
\usage{
dna_version_id(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Further arguments passed to methods}
}
\description{
Extract the DNA Version ID
}
\keyword{internal}

