% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inject-errors.r
\name{do_inject}
\alias{do_inject}
\title{Inject errors}
\usage{
do_inject(dt, args)
}
\arguments{
\item{dt}{A data.table with genotypes and error probabilities}

\item{args}{A list containing a mapper function <map>, a
remapper function <remap>, and <odds>.}
}
\description{
Inject errors
}
\keyword{internal}

