% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allele.r
\name{eag_allele}
\alias{eag_allele}
\title{Class: eag_allele}
\usage{
eag_allele(a1, a2)
}
\arguments{
\item{a1, }{a2 Alleles}
}
\description{
Constructor for an <eag_allele>.
}
\details{
Inherits from <list>.
}
\seealso{
\code{\link{eag_numbers}},  \code{\link{rep_allele}}
}
\keyword{internal}

