% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allele.r
\name{eag_numbers}
\alias{eag_numbers}
\title{Class: eag_numbers}
\usage{
eag_numbers(e2, e3)
}
\arguments{
\item{e2, }{e3 EAG numbers for exon 2 and 3.}
}
\description{
Constructor for <eag_numbers>.
}
\seealso{
\code{\link{eag_allele}},  \code{\link{rep_allele}}
}
\keyword{internal}

