% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inject-errors.r
\name{eag_status}
\alias{eag_status}
\title{How many eags do we find on exons 2 and 3?}
\usage{
eag_status(nums)
}
\arguments{
\item{nums}{Eag number}
}
\description{
11 <-> hom:hom
12 <-> hom:het
21 <-> het:hom
22 <-> het:het
}
\keyword{internal}

