% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tables.r
\name{eag_table}
\alias{eag_table}
\title{Class: eag_tbl}
\usage{
eag_table(gene = "DPB1", nextype_basis_id = "1412")
}
\arguments{
\item{gene}{An HLA gene. One of \sQuote{A}, \sQuote{B}, \sQuote{C}, \sQuote{DRB1},
\sQuote{DQB1}, \sQuote{DPB1}.}

\item{nextype_basis_id}{"NeXtype Basis ID".}
}
\value{
A table with the fields:
\itemize{
 \item "eag_num": Numeric code describing an Exon Allele Group.
 \item "eag_allele": HLA allele code.
 \item "exon": Exon '2' or '3'.
 \item "allele_id": Allele identifier.
}
and the attributes:
\itemize{
 \item "nextype_basis_id"
 \item "gene"
}
}
\description{
Constructor for an <\code{eag_tbl}> object.

THIS FUNCTION REQURIES ACCESS TO INTERNAL DATABASES AT DKMS LSL.

Use the prepackaged datasets (\code{\link{eag_tables}}) instead.
}
\examples{
\dontrun{
dpb1_eag1412 <- eag_table("DPB1", "1412")
}
}

