% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{eag_tables}
\alias{dpb1_eag1412}
\alias{dqb1_eag1412}
\alias{drb1_eag1412}
\alias{eag_tables}
\title{EAG tables}
\format{Data tables with 4 columns:
\describe{
 \item{eag_num}{Unique numeric code for an Exon Allele Group.}
 \item{eag_allele}{HLA allele code}
 \item{exon}{Exon "2" or "3"}
 \item{allele_id}{Allele identifier}
}}
\usage{
drb1_eag1412

dqb1_eag1412

dpb1_eag1412
}
\description{
Datasets containing the EAG numbers and the corresponding alleles
for DRB1, DQB1, and DPB1 based on NeXtype Basis ID 1412 for Exons
2 and 3
}
\examples{
drb1_eag1412
}
\keyword{datasets}

