% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.r
\name{encode_ambiguities}
\alias{encode_ambiguities}
\title{Encode subtypes to G or NMDP codes}
\usage{
encode_ambiguities(x, gtbl, ntbl2)
}
\arguments{
\item{x}{A vector of alleles}
}
\description{
Encode subtypes to G or NMDP codes
}
\details{
Expect gtbl and ntbl2 (with 2-field subtypes)
in the parent environment.
}
\keyword{internal}

