% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.r
\name{errors}
\alias{errors}
\title{Extract errors}
\usage{
errors(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Further arguments passed to methods}
}
\description{
Extract errors
}
\keyword{internal}

