% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generics.r
\name{exon}
\alias{exon}
\title{Extract an exon}
\usage{
exon(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Further arguments passed to methods}
}
\description{
Extract an exon
}
\keyword{internal}

