% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/expected-genotype-frequency.r
\name{expected_genotype_frequencies}
\alias{expected_genotype_frequencies}
\title{Calculate expected genotype frequencies.}
\usage{
expected_genotype_frequencies(alleles, frequencies)
}
\arguments{
\item{alleles}{A character vector of alleles.}

\item{frequencies}{The corresponding allele frequencies.}
}
\value{
A data table with two columns:
\itemize{
 \item{genotype}
 \item{pexp}
}
}
\description{
Calculate expected genotype frequencies.
}
\keyword{internal}

