% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generics.r
\name{gene}
\alias{gene}
\title{Extract the gene name}
\usage{
gene(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Further arguments passed to methods}
}
\description{
Extract the gene name
}
\keyword{internal}

