% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{geno_data}
\alias{dqb1}
\alias{drb1}
\alias{geno_data}
\title{Genotyping results for class II HLA genes.}
\format{\itemize{
  \item \code{drb1}: A <\code{\link{HLA}}> instance with 926,228 rows.
  \item \code{dqb1}: A <\code{\link{HLA}}> instance with 922,495 rows.
  \item \code{dpb1}: A <\code{\link{HLA}}> instance with 900,913 rows.
}}
\usage{
drb1

dqb1

dqb1
}
\description{
Genotyping results for the class II HLA genes, DRB1, DQB1, and DPB1,
processed between 01/01/2014 and 23/03/2015 at DKMS Life Science Lab.
}
\examples{
drb1
drb1$get_table()
drb1.de <- drb1[provenance == "DE"]
drb1.de$allele_frequency()
drb1.de$genotype_frequency()
}
\keyword{datasets}

