% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/testing.r
\name{get_testdata}
\alias{get_testdata}
\title{Retrieve testing data from ngsrep for a gene}
\usage{
get_testdata(gene, nextype_basis_id)
}
\arguments{
\item{gene}{A HLA gene.}

\item{nextype_basis_id}{A neXtype Basis ID.}
}
\value{
A table with the fields:
\describe{
 \item{lims_donor_id}{}
 \item{nextype_run_id}{}
 \item{orig_nmdp1}{}
 \item{orig_nmdp2}{}
 \item{exon}{}
 \item{eag_num}{}
 \item{main_eag}{}
 \item{total_reads}{}
 \item{best_reads}{}
}
}
\description{
Retrieve testing data from ngsrep for a gene
}
\examples{
\dontrun{
testdata <- get_testdata("DPB1", "1267")
}
}
\keyword{internal}

