% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tables.r
\name{gtf_table.HLA}
\alias{gtf_table.HLA}
\title{Class: gtf_tbl}
\usage{
\method{gtf_table}{HLA}(x)
}
\arguments{
\item{x}{A <\code{\link{HLA}}> object.}
}
\value{
A table with the fields:
\itemize{
 \item "genotype": \strong{Key}. The genotype in the format "A1/A2".
 \item "pexp": Expected genotype frequencies.
 \item "pobs": Observed genotype frequencies.
 \item "log_pexp": Log of expected genotype frequenciies.
 \item "log_pobs": Log of expected genotype frequenciies.
 \item "nexp": Expected number of samples for a genotype.
 \item "nobs": Observed number of samples for a genotype.
 \item "chisq": \eqn{\chi^2}-value describing the difference
   between expected and observed genotype frequency.
 \item "log_fold_diff": Log-fold diffence between expected
   and observed genotype frequency.
}
}
\description{
Constructor for <\code{gtf_tbl}> objects. Calculates observed
and expected genotype frequencies.
}
\examples{
\dontrun{
## Extract HLA-DPB1 genotype frequencies
dpb1 <- HLA("DPB1", "01/01/2014", "23/03/2015")

## Restrict the data to the German sample
dpb1.de <- dpb1[provenance == "DE"]

## Remove low-resolution four-digit codes
dpb1.de <- dpb1.de[allele1 \%ni\% dpb1_four_digit_codes | allele2 \%ni\% dpb1_four_digit_codes]

## Remove cases of unknown alleles
dpb1.de <- dpb1.de[field2(allele1) != "XXX" | field2(allele2) != "XXX"]
dpb1.de <- dpb1.de[toupper(allele1) != "NEW" | toupper(allele2) != "NEW"]

## Calculate observed and expected genotype frequencies
gtf_table(dpb1.de)
}
}
\keyword{internal}

