% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.r
\name{gtf_table}
\alias{gtf_table}
\title{Extract table of genotype frequencies}
\usage{
gtf_table(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Further arguments passed to methods}
}
\description{
Extract table of genotype frequencies
}
\keyword{internal}

