% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hla.r
\name{HLA}
\alias{HLA}
\title{Genotyping data generated at DKMS Life Science Lab.}
\usage{
HLA(gene = NULL, from = NULL, to = NULL)
}
\arguments{
\item{gene}{Fetch alleles for one of "A", "C", "B", "DRB1", "DQB1", "DPB1".}

\item{from}{Start query at DD/MM/YYYY.}

\item{to}{End query at DD/MM/YYYY.}
}
\value{
A <HLA> object containing a table with the following fields:
\describe{
   \item{lims_donor_id}{Unique sample tracking ID.}
   \item{provenance}{One of \emph{DE}, \emph{PL}, or \emph{UK}.}
   \item{allele1}{First allele.}
   \item{allele2}{Other allele.}
   \item{genotype}{The genotype in the format "allele1/allele2", where
   the alleles are alphabetically ordered.}
   \item{zygosity}{One of \emph{homozygous} or \emph{heterozygous}}
}
}
\description{
An R6 reference class holding genotyping results for any of the
six HLA genes routinely typed at DKMS LSL.

THIS FUNCTION REQURIES ACCESS TO INTERNAL DATABASES AT DKMS LSL.

Use the prepackaged datasets (\code{\link{geno_data}}) instead.
}
\examples{
\dontrun{
dpb1 <- HLA("DPB1", "01/01/2014", "23/04/2015")

# restrict the sample to a country of origin
dpb1.de <- dpb1[provenance == "DE"]

# access the data table
dpb1.de.tbl <- dpb1.de$get_table()

# calculate allele frequencies
dpb1.de.af <- dpb1.de$allele_frequency()

# calculate genotype frequencies
dpb1.de.gf <- dpb1.de$genotype_frequency()
}
}
\keyword{internal}

