% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/inject-errors.r
\name{inject_errors}
\alias{inject_errors}
\title{Inject errors into a sample of genotypes.}
\usage{
inject_errors(x, cdf, perr, odds = 1)
}
\arguments{
\item{x}{A \code{\link{geno_table}} object with attribute \code{has_errors = FALSE}.}

\item{cdf}{A cumulative distribution function describing a DNA-concentration-
dependent error rate. See \code{\link{cdfA}} and \code{\link{cdfB}} for two
precomputed error functions based on HLA-A and HLA-B.}

\item{perr}{The overall error rate.}

\item{odds}{The odds of loosing an EAG on exon 2 or exon 3, respectively.
E.g. \code{odds = 2} means a 2:1 chance that the loss occurs on Exon 2;
\code{odds = 0.25} means a 1:4 chance that the loss occurs on Exon 3.}
}
\value{
A \code{\link{geno_table}} object with attribute \code{has_errors = TRUE}.
}
\description{
Takes a \code{\link{geno_table}} object and injects a total of
\emph{perr} errors with a concentration-dependent rate described
by a cumulative distribution functions, \emph{cdf}, using the
probability calculation described in \code{\link{loose_eag}}..
}
\note{
The next step is to map genotype frequencies onto the sampled
  genotypes: \code{\link{map_genotype_frequencies}}.
}
\examples{
\dontrun{
## Extract HLA-DPB1 genotype frequencies
dpb1 <- HLA("DPB1", "01/01/2014", "23/03/2015")

## Restrict the data to the German sample
dpb1.de <- dpb1[provenance == "DE"]

## Generate an EAG table
dpb1_eag1412 <- eag_table(gene = "DPB1", nextype_basis_id = "1412")

## Generate a distribution of DNA concentrations
conc <- sample_dna_concentration(dpb1.de, n = 1000, ncores = 8)

## Generate a sampling function
sample_dpb1_de <- make_genotype_sampler(dpb1.de, eag_1267)

## Load precomputed DNA-concentration-dependent error distribution for HLA-A
data(cdfA)

## Sample genotypes
n <- 10000
bin_size <- 3
perr <- 0.01
odds <- 0.25
ans <- sample_dpb1_de(conc, n, bin_size) \%>\%
  inject_errors(cdfA, perr, odds)
ans

summary(ans)
samples(ans)
errors(ans)
merge(ans)
}
}
\seealso{
Other simulation.functions: \code{\link{make_genotype_sampler}};
  \code{\link{map_genotype_frequencies}}
}

