% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.r
\name{lookup_list}
\alias{lookup_list}
\title{Class: lookup_list}
\usage{
lookup_list(alleles, eag)
}
\arguments{
\item{alleles}{A character vector of alleles.}

\item{eag}{EAG alleles <\code{\link{eag_tbl}}>.}
}
\value{
A list with the elements:
\describe{
 \item{e2_tbl}{A mapping of rep_allele <-> eag_alllele <-> eag_num for exon 2}
 \item{e3_tbl}{A mapping of rep_allele <-> eag_alllele <-> eag_num for exon 2}
 \item{jkr_tbl}{A table of exon 3 joker alleles}
 \item{prt_tbl}{A mapping table with partials}
 \item{ntbl2}{2-field NMDP-code lookup}
 \item{ntbl4}{4-field NMDP-code lookup}
 \item{gtbl}{G-code lookup}
}
and the attributes:
\itemize{
 \item{nextype_basis_id}
 \item{gene}
}
}
\description{
Constructor for a <lookup_list> object.
}
\details{
Inherits from <list>.
}
\examples{
\dontrun{
library("HLA")

## get a vector of reported DPB1 alleles
dpb1 <- HLA::HLA("DPB1")
alleles <- HLA::get_table(dpb1)[, .(allele1, allele2)]
alleles <- sort(unique(c(alleles[, allele1], alleles[, allele2])))
alleles <- alleles[-grep("(:XXX|NEW)", alleles)]

## get an EAG table for DPB1 and a NeXtype Basis ID
eag <- eag_table("DPB1", "1267")

lookup <- lookup_list(alleles, eag)
}
}

