% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/inject-errors.r
\name{loose_eag}
\alias{loose_eag}
\title{Loose an EAG}
\usage{
loose_eag(x, pE, odds = 1)
}
\arguments{
\item{x}{A <\code{\link{eag_numbers}}> instance.}

\item{pE}{The absolute probability that an error occurs.}

\item{odds}{The odds in favour of loosing an EAG on Exon 2 rather
than Exon 3. E.g. \code{odds = 2} means a 2:1 chance that the loss
occurs on Exon 2; \code{odds = 0.25} means a 1:4 chance that
the loss occurs on Exon 3.}
}
\value{
An <eag_numbers> instance
}
\description{
Randomly loose an EAG number.
}
\details{
We generally deal with the following scenario on any HLA locus currently:

\tabular{rcc}{
        \tab Exon A \tab Exon B \cr
  EAG 1 \tab a      \tab c      \cr
  EAG 2 \tab b      \tab d
}

We know (have a guesstimate of) the total error rate due to loss of EAGs
at a given DNA concentration, \eqn{P(E)}, and we know the odds, \eqn{\omega},
that we'll loose an EAG at one or the other exon.

E.g. if we perform two independent PCR reactions for exon 2 and only one
PCR reaction for exon 3, the odds of loosing an EAG are 4 in 1 for exon
3. I. e. in order for an error to manifest on exon 2 we must loose the
same EAG in both reactions, the probability of which is \out{&frac14;}
of loosing an EAG in a single reaction.

What we want to know is:
\itemize{
  \item \out{P(A &cap; &not; B | E)}
  The probability that we loose an EAG on exon A but not on exon B,
  given that we make an error at all.
  \item \out{P(&not; A &cap; B | E)}
  The probability that we loose an EAG on exon B but not on exon A,
  given that we make an error at all.
  \item \out{P(A &cap; B | E)}
  The probability that we loose an EAG on exon A and on exon B,
  given that we make an error at all.
}

To get there we need to know the error rate for exon A, \out{&epsilon;<sub>A</sub>},
and the error rate for exon B, \out{&epsilon;<sub>B</sub>}, since:
\itemize{
 \item \out{P(A &cap; &not; B; | E) = &epsilon;<sub>A</sub>(1 - &epsilon;<sub>B</sub>)
                                      &frasl;
                                      (&epsilon;<sub>A</sub> + &epsilon;<sub>B</sub> -
                                      &epsilon;<sub>A</sub>&epsilon;<sub>B</sub>)}
 \item \out{P(&not; A &cap; B; | E) = &epsilon;<sub>B</sub>(1 - &epsilon;<sub>A</sub>)
                                      &frasl;
                                      (&epsilon;<sub>A</sub> + &epsilon;<sub>B</sub> -
                                      &epsilon;<sub>A</sub>&epsilon;<sub>B</sub>)}
 \item \out{P(&not; A &cap; B; | E) = &epsilon;<sub>B</sub>&epsilon;<sub>A</sub>
                                      &frasl;
                                      (&epsilon;<sub>A</sub> + &epsilon;<sub>B</sub> -
                                      &epsilon;<sub>A</sub>&epsilon;<sub>B</sub>)}
}

Considering that \out{&epsilon;<sub>A</sub> + &epsilon;<sub>B</sub> -
&epsilon;<sub>A</sub>&epsilon;<sub>B</sub> = P(E)} and
\out{&epsilon;<sub>B</sub> = <sup>1</sup>&frasl;<sub>&omega;</sub>
&epsilon;<sub>A</sub>} we can arrive after some rearrangements at:

\itemize{
 \item \out{&epsilon;<sub>A</sub> = &frac12;(1 + &omega; - &omega;
 &radic;<span style="text-decoration: overline">
 (1 + 1&frasl;&omega;)<sup>2</sup> -
 4<em>P(E)</em>&frasl;&omega;</span>)}
 \item \out{&epsilon;<sub>B</sub> = <sup>1</sup>&frasl;<sub>&omega;</sub>
&epsilon;<sub>A</sub>}
 }

 These error rates allow now to randomly loose EAGs with the
 correct probabilities.
}
\keyword{internal}

