% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mapping.r
\name{make_encoder}
\alias{make_encoder}
\title{Make a function that encodes eag alleles}
\usage{
make_encoder(lookup)
}
\arguments{
\item{lookup}{}
}
\value{
A function that takes an <eag_allele> as input
and returns a <rep_allele>
}
\description{
Make a function that encodes eag alleles
}
\keyword{internal}

