% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geno-table.r
\name{make_genotype_sampler}
\alias{make_genotype_sampler}
\title{Generate a genotype resampling function.}
\usage{
make_genotype_sampler(x, eag)
}
\arguments{
\item{x}{A <\code{\link{HLA}}> object.}

\item{eag}{An <\code{\link{eag_table}}> object.}
}
\value{
A constructor function for <\code{\link{geno_table}}> objects.
}
\description{
Generate a genotype resampling function.
}
\note{
The next step is to inject errors in the sampled genotypes:
\code{\link{inject_errors}}.
}
\examples{
\dontrun{
## Extract HLA-DPB1 genotype frequencies
dpb1 <- HLA("DPB1", "01/01/2014", "23/03/2015")

## Restrict the data to the German sample
dpb1.de <- dpb1[provenance == "DE"]

## Generate an EAG table
dpb1_eag1412 <- eag_table(gene = "DPB1", nextype_basis_id = "1412")

## Sample a distribution of DNA concentrations
concentration <- sample_dna_concentration(dpb1.de, n = 24000, ncores = 8)

## Generate a sampling function
sample_dpb1_de <- make_genotype_sampler(dpb1.de, dpb1_eag1412)

## Sample genotypes
n <- 10000
bin_size <- 3
rs <- sample_dpb1_de(concentration, n, bin_size)
rs
}
}
\seealso{
Other simulation.functions: \code{\link{inject_errors}};
  \code{\link{map_genotype_frequencies}}
}

