% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mapper.r, R/remapper.r
\name{make_mapper}
\alias{make_mapper}
\alias{make_remapper}
\title{Generate a mapping function}
\usage{
make_mapper(lookup)

make_remapper(lookup)
}
\arguments{
\item{lookup}{A \code{\link{lookup_list}} object.}
}
\value{
A function
}
\description{
Make mappers for mapping \code{EAG numbers -> Report Alleles} (\code{make_mapper}) and
for mapping \code{Report Alleles -> EAG numbers} (\code{make_remapper}).
}
\examples{
## load the prepackaged date for DPB1 report alleles and
## the DPB1 EAG table
data(rep_dpb1, package = "HLAdata"); data(eag, package = "HLAdata")

lookup <- lookup_list(rep_dpb1, eag)
map <- make_mapper(lookup)

nums <- eag_numbers(e2 = c("6680148", "6673401"), e3 = "6670791")
map(nums)

remap <- make_remapper(lookup)
new_nums <- remap(map(nums))

nums == new_nums
}

