% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/testing.r
\name{make_test_instance}
\alias{make_test_instance}
\alias{make_test_instance2}
\title{Make a tester}
\usage{
make_test_instance(testdata, lookup)

make_test_instance2(alleles, mapper, remapper)
}
\arguments{
\item{testdata}{A dataset generated by a call to \code{\link{get_testdata}}.}

\item{lookup}{A <\code{\link{lookup_table}}>.}

\item{alleles}{A vector of alleles}

\item{mapper}{A <mapper> function.}

\item{remapper}{A <remapper> fucntion.}
}
\value{
A list of functions.
}
\description{
Generate a list of functions \code{$test()}, \code{$get_successes(all = FALSE)},
\code{$get_failures(all = FALSE)}, \code{$reset()}, \code{$get_run_ldids()},
that allow testing the performance of the mapper against NeXtype results fro
a test dataset.
}
\examples{
lookup <- lookup_list(rep_dpb1, eag)
tester <- make_test_instance(testdata, lookup)
tester$test(limit = 20)
tester$get_successes()
tester$get_failures()
tester$test(limit = 20)
tester$get_failures()

run_tests_on("ID11732060", testdata, lookup)
}
\keyword{internal}

