% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mapping.r
\name{map_rep_allele}
\alias{map_rep_allele}
\title{Map an eag allele to a report allele}
\usage{
map_rep_allele(a, lookup)
}
\arguments{
\item{a}{An <\code{\link{eag_allele}}>.}

\item{lookup}{A <\code{\link{lookup_list}}>}
}
\value{
A <\code{\link{rep_allele}}>.
}
\description{
Map an eag allele to a report allele
}
\examples{
nums <- eag_numbers(e2 = c(6673401, 6673404), e3 = c(6672138, 6670791))
lookup <- lookup_list(rep_dpb1, eag)

a <- map_eag_allele(nums, lookup)
rp <- map_rep_allele(a, lookup)

rp1 <- rep_allele(a1 = "04:ADCGE", a2 = "13:01:01G")
rp == rp1
}
\keyword{internal}

