% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.r
\name{mapper}
\alias{mapper}
\title{Extract mapper function}
\usage{
mapper(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Further arguments passed to methods}
}
\description{
Extract mapper function
}
\keyword{internal}

