% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generics.r
\name{nextype_basis_id}
\alias{nextype_basis_id}
\title{Extract the NeXtype Basis ID}
\usage{
nextype_basis_id(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Further arguments passed to methods}
}
\description{
Extract the NeXtype Basis ID
}
\keyword{internal}

