% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generics.r
\name{remapper}
\alias{remapper}
\title{Extract remapper function}
\usage{
remapper(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Further arguments passed to methods}
}
\description{
Extract remapper function
}
\keyword{internal}

