% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/allele.r
\name{rep_allele}
\alias{!=.rep_allele}
\alias{==.rep_allele}
\alias{allele.rep_allele}
\alias{is.na.rep_allele}
\alias{print.rep_allele}
\alias{rep_allele}
\title{Class: rep_allele}
\usage{
rep_allele(a1 = NA_character_, a2 = NA_character_)

\method{print}{rep_allele}(x, ...)

\method{is.na}{rep_allele}(x)

\method{==}{rep_allele}(e1, e2)

\method{!=}{rep_allele}(e1, e2)

\method{allele}{rep_allele}(x, n)
}
\arguments{
\item{a1,}{a2 Alleles}
}
\description{
Constructor for a <rep_allele>.
}
\details{
Inherits from <list>.
}
\section{Methods (by generic)}{
\itemize{
\item \code{print}: 

\item \code{is.na}: 

\item \code{==}: 

\item \code{!=}: 

\item \code{allele}: 
}}
\seealso{
\code{\link{eag_numbers}},  \code{\link{eag_allele}}
}
\keyword{internal}

