% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna-concentration.r
\name{sample_dna_concentration}
\alias{sample_dna_concentration}
\title{Extract a random sample of DNA concentrations.}
\usage{
sample_dna_concentration(x, n = 10000, ncores = 4)
}
\arguments{
\item{x}{A <\code{\link{HLA}}> object.}

\item{n}{Number of samples.}

\item{ncores}{How many cores to we use for querying the database.}
}
\value{
A \code{data.table} with the fields:
\itemize{
 \item "lims_donor_id": LIMS_DONOR_ID
 \item "ymd": A timestamp.
 \item "conc": The DNA concentration in ng/µl.
}
}
\description{
THIS FUNCTION REQURIES ACCESS TO INTERNAL DATABASES AT DKMS LSL.

Use the prepackaged dataset (\code{\link{concentration}}) instead.
}
\examples{
\dontrun{
## Generate a random distribution of DNA concentrations
data("dpb1")
conc <- sample_dna_concentration(dpb1, n = 12000, ncores = 12)
}
}

