% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generics.r
\name{samples}
\alias{samples}
\title{Extract samples}
\usage{
samples(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Further arguments passed to methods}
}
\description{
Extract samples
}
\keyword{internal}

