% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generics.r
\name{allele}
\alias{allele}
\title{Extract an allele}
\usage{
allele(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Further arguments passed to methods}
}
\description{
Extract an allele
}
\keyword{internal}

