% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/allele.r
\name{eag_allele}
\alias{!=.eag_allele}
\alias{==.eag_allele}
\alias{allele.eag_allele}
\alias{eag_allele}
\alias{is.na.eag_allele}
\alias{print.eag_allele}
\title{Class: eag_allele}
\usage{
eag_allele(a1, a2)

\method{is.na}{eag_allele}(x)

\method{print}{eag_allele}(x, ...)

\method{==}{eag_allele}(e1, e2)

\method{!=}{eag_allele}(e1, e2)

\method{allele}{eag_allele}(x, n)
}
\arguments{
\item{a1,}{a2 Alleles}
}
\description{
Constructor for an <eag_allele>.
}
\details{
Inherits from <list>.
}
\section{Methods (by generic)}{
\itemize{
\item \code{is.na}: 

\item \code{print}: 

\item \code{==}: 

\item \code{!=}: 

\item \code{allele}: 
}}
\seealso{
\code{\link{eag_numbers}},  \code{\link{rep_allele}}
}
\keyword{internal}

