% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/allele.r
\name{eag_numbers}
\alias{!=.eag_numbers}
\alias{==.eag_numbers}
\alias{eag_numbers}
\alias{exon.eag_numbers}
\alias{is.na.eag_numbers}
\alias{print.eag_numbers}
\title{Class: eag_numbers}
\usage{
eag_numbers(e2, e3)

\method{print}{eag_numbers}(x)

\method{exon}{eag_numbers}(x, n)

\method{==}{eag_numbers}(e1, e2)

\method{!=}{eag_numbers}(e1, e2)

\method{is.na}{eag_numbers}(x)
}
\arguments{
\item{e2,}{e3 EAG numbers for exon 2 and 3.}
}
\description{
Constructor for <eag_numbers>.
}
\section{Methods (by generic)}{
\itemize{
\item \code{print}: 

\item \code{exon}: 

\item \code{==}: 

\item \code{!=}: 

\item \code{is.na}: 
}}
\seealso{
\code{\link{eag_allele}},  \code{\link{rep_allele}}
}
\keyword{internal}

