% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.r
\name{expand_ambiguity_codes}
\alias{expand_ambiguity_codes}
\title{Expand ambiguity codes}
\usage{
expand_ambiguity_codes(alleles, lookup)
}
\description{
Input: a vector of alleles with ambiguity codes (NMDP and G)
Output: a vector of alleles with ambiguity codes expanded.
}
\keyword{internal}

