% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lookup.r
\name{lookup_list}
\alias{exon.lookup_list}
\alias{gene.lookup_list}
\alias{lookup_list}
\alias{nextype_basis_id.lookup_list}
\alias{print.lookup_list}
\title{Class: lookup_list}
\usage{
lookup_list(alleles, eag)

\method{print}{lookup_list}(x, ..., n = 2)

\method{gene}{lookup_list}(x)

\method{nextype_basis_id}{lookup_list}(x)

\method{exon}{lookup_list}(x, n)
}
\arguments{
\item{alleles}{A character vector of alleles.}

\item{eag}{EAG alleles <\code{\link{eag_tbl}}>.}
}
\value{
A list with the elements:
\describe{
 \item{e2_tbl}{A mapping of rep_allele <-> eag_alllele <-> eag_num for exon 2}
 \item{e3_tbl}{A mapping of rep_allele <-> eag_alllele <-> eag_num for exon 2}
 \item{jkr_tbl}{A table of exon 3 joker alleles}
 \item{prt_tbl}{A mapping table with partials}
 \item{ntbl2}{2-field NMDP-code lookup}
 \item{ntbl4}{4-field NMDP-code lookup}
 \item{gtbl}{G-code lookup}
}
and the attributes:
\itemize{
 \item{nextype_basis_id}
 \item{gene}
}
}
\description{
Constructor for a <lookup_list> object.
}
\details{
Inherits from <list>.
}
\section{Methods (by generic)}{
\itemize{
\item \code{print}: 

\item \code{gene}: 

\item \code{nextype_basis_id}: 

\item \code{exon}: 
}}
\examples{
\dontrun{
library("HLA")

## get a vector of reported DPB1 alleles
dpb1 <- HLA::HLA("DPB1")
alleles <- HLA::get_table(dpb1)[, .(allele1, allele2)]
alleles <- sort(unique(c(alleles[, allele1], alleles[, allele2])))
alleles <- alleles[-grep("(:XXX|NEW)", alleles)]

## get an EAG table for DPB1 and a NeXtype Basis ID
eag <- eag_table("DPB1", "1267")

lookup <- lookup_list(alleles, eag)
}
}
\keyword{internal}

