% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.r
\name{lookup_table}
\alias{lookup_table}
\title{Class: lookup_table}
\usage{
lookup_table(allele_tbl, eagn)
}
\arguments{
\item{eagn}{An \code{\link{eag_tbl}} for either Exon 2 or 3 .}
}
\value{
A table with the fields:
\describe{
 \item{rep_allele}{The reported allele codes (including NMDP and G codes)}
 \item{eag_allele}{The fully resolved HLA allele codes}
 \item{eag_num}{Unique numeric code for an Exon Allele Group.}
}
and the attributes:
\itemize{
 \item{nextype_basis_id}
 \item{gene}
 \item{exon}
}
}
\description{
Constructor for a <\code{lookup_table}>.
}
\details{
Inherits from <\code{data.table}>
}
\seealso{
\code{\link{lookup_list}}
}

