% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.r
\name{make_disambiguator}
\alias{make_disambiguator}
\title{Make a function that disambiguates report alleles}
\usage{
make_disambiguator(lookup)
}
\value{
A function that takes a <rep_allele> as input
and returns an <eag_allele>
}
\description{
Make a function that disambiguates report alleles
}
\keyword{internal}

