% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allele.r
\name{rep_allele}
\alias{rep_allele}
\title{Class: rep_allele}
\usage{
rep_allele(a1 = NA_character_, a2 = NA_character_)
}
\arguments{
\item{a1, }{a2 Alleles}
}
\description{
Constructor for a <rep_allele>.
}
\details{
Inherits from <list>.
}
\seealso{
\code{\link{eag_numbers}},  \code{\link{eag_allele}}
}
\keyword{internal}

